% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{anno_matrix}
\alias{anno_matrix}
\title{IlluminaHumanMethylation450kanno}
\format{
IlluminaHumanMethylation450kanno.ilmn12.hg19 annotation file.
This data have several columns
}
\usage{
data(anno_matrix)
}
\description{
IlluminaHumanMethylation450kanno
}
\keyword{datasets}
