% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.ridgemat}
\alias{as.data.frame.ridgemat}
\title{data frame ridge matrix}
\usage{
\method{as.data.frame}{ridgemat}(x, ...)
}
\arguments{
\item{x}{data vector}

\item{...}{other parameters pass to as.data.frame.model.matrix()}
}
\description{
data frame ridge matrix
}
