% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{demo_MCBinformation}
\alias{demo_MCBinformation}
\title{MCB information.}
\format{A data frame with 94 rows and 8 variables:
\describe{
  \item{MCB_no}{MCB code}
  \item{start}{Start point of this MCB in the chromosome.}
  \item{end}{End point of this MCB in the chromosome.}
  \item{CpGs}{All the CpGs probe names in the MCB.}
  \item{location}{Start, end point and the chromosome number of this MCB.}
  \item{chromosomes}{the chromosome number of this MCB.}
  \item{length}{the length of bps of this MCB in the chromosome.}
  \item{CpGs_num}{number of CpG probes of this MCB.}
}}
\usage{
data(demo_MCBinformation)
}
\description{
A dataset containing the number and other attributes of 94 MCBs; 
This results was created by the identification function IdentifyMCB.
This data used for metricMCB function.
}
\keyword{datasets}
