% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{demo_set}
\alias{demo_set}
\title{ExpressionSet of demo dataset.}
\format{ExpressionSet:
\describe{
  \item{assays}{10020 features, 455 samples}
  \item{colData}{Data descriptions.}
  ...
}}
\usage{
data(demo_set)
}
\description{
A ExpressionSet containing the 10020 CpGs beta value of 455 samples.
}
\keyword{datasets}
