% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{demo_survival_data}
\alias{demo_survival_data}
\title{Survival data of demo dataset.}
\format{Surv data:}
\usage{
data(demo_survival_data)
}
\description{
A Surv containing the 1020 CpGs beta value of 455 samples.
}
\keyword{datasets}
