% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw_survival_curve}
\alias{draw_survival_curve}
\title{draw survival curve}
\usage{
draw_survival_curve(
  exp,
  living_days,
  living_events,
  write_name,
  title_name = "",
  threshold = NA
)
}
\arguments{
\item{exp}{expression level for gene.}

\item{living_days}{The survival time (days) for each individual.}

\item{living_events}{The survival event for each individual, 0 indicates alive and 1 indicates death.
Other choices are TRUE/FALSE (TRUE = death) or 1/2 (2=death). For interval censored data, the status indicator is 0=right censored, 1=event at time, 2=left censored, 3=interval censored.}

\item{write_name}{The name for pdf file which contains the result figure.}

\item{title_name}{The title for the result figure.}

\item{threshold}{Threshold used to indicate the high risk or low risk.}
}
\value{
This function will generate a pdf file with 300dpi which compare survival curves using the Kaplan-Meier (KM) test.
}
\description{
Draw a survival curve based on survminer package. This is a wrapper function of ggsurvplot.
}
\author{
Xin Yu
}
