% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_prediction.R
\name{ensemble_prediction}
\alias{ensemble_prediction}
\title{fitting function using stacking ensemble model for Methylation Correlation Block}
\usage{
ensemble_prediction(ensemble_model, predition_data)
}
\arguments{
\item{ensemble_model}{ensemble model which built by ensemble_model() function}

\item{predition_data}{A vector, matrix, list, or data frame containing the predictions (input).}
}
\value{
Object of numeric class \code{double}
}
\description{
predict is a generic function for predictions from the results of stacking ensemble model fitting functions.
The function invokes particular methods which is the ensemble model described in the reference.
}
\references{
Xin Yu et al. 2019 Predicting disease progression in lung adenocarcinoma patients based on methylation correlated blocks using ensemble machine learning classifiers (under review)
}
