% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyMCB_parallel.R
\name{IdentifyMCB_parallel}
\alias{IdentifyMCB_parallel}
\title{Identification of methylation correlated blocks using parallel}
\usage{
IdentifyMCB_parallel(MethylationProfile, method = c("pearson",
  "spearman", "kendall")[1], CorrelationThreshold = 0.8,
  PositionGap = 1000, platform = "Illumina Methylation 450K")
}
\arguments{
\item{MethylationProfile}{Methylation matrix is used in the analysis.}

\item{method}{method used for calculation of correlation, should be one of "pearson","spearman","kendall". Defualt is "pearson".}

\item{CorrelationThreshold}{Pearson correlation threshold is used for define boundaries.}

\item{PositionGap}{CpG Gap between any two CpGs positioned CpG sites less than 1000 bp (default) will be calculated.}

\item{platform}{This parameter indicates the platform used to produce the methlyation profile.}
}
\value{
Object of class \code{list} with elements:
 \tabular{ll}{
   \code{MCBsites} \tab Character set contains all CpG sites in MCBs. \cr
   \code{MCBinformation} \tab Matrix contains the information of results. \cr
 }
}
\description{
This function is simiar to the IdentifyMCB function, but designed for parallel calculation,
This package partition the genome into blocks of tightly co-methylated CpG sites, namely Methylation correlated blocks. This function calculates Pearson correlation coefficients r^2 between
the beta values of any two CpGs r^2 < CorrelationThreshold was used to identify boundaries between any two
adjacent markers indicating uncorrelated methylation. Markers not separated by a boundary were combined into MCB. Pearson correlation coefficients between
two adjacent CpGs were calculated.
}
\details{
Currently, only illumina 450k platform is supported, the methylation profile need to convert into ExpressionSet format.
Task parallelism strategy is used for acceleration of calculation. Since this strategy is memory consuming, this function is strongly recommended to run on a server.
}
\seealso{
IdentifyMCB
}
\author{
Xin Yu
}
