% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predict.mcb.coxph.penal}
\alias{predict.mcb.coxph.penal}
\title{predict coxph penal using MCB}
\usage{
\method{predict}{mcb.coxph.penal}(object, newdata, ...)
}
\arguments{
\item{object}{the results of a coxph fit.}

\item{newdata}{Optional new data at which to do predictions. If absent predictions are for the data frame used in the original fit. 
When coxph has been called with a formula argument created in another context, i.e., coxph has been called within another function and the formula was passed as an argument to that function, 
there can be problems finding the data set. See the note below.}

\item{...}{other parameters pass to predict.coxph}
}
\value{
prediction values of regression.
}
\description{
Compute fitted values and regression terms for a model fitted by coxph
}
\author{
Xin Yu
}
