% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_demo}
\alias{create_demo}
\title{create demo matrix}
\usage{
create_demo()
}
\value{
This function will generate a demo data.
}
\description{
Demo matrix for methylation matrix.
}
\examples{
demo_set<-create_demo()
}
\author{
Xin Yu
}
