% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{demo_MCBinformation}
\alias{demo_MCBinformation}
\title{MCB information.}
\format{A data frame with 94 rows and 8 variables:
\describe{
  \item{MCB_no}{MCB code}
  \item{start}{start point of this MCB in the chromosome.}
  ...
}}
\usage{
data(demo_MCBinformation)
}
\description{
A dataset containing the number and other attributes of 94
MCBs.
}
\keyword{datasets}
