% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{demo_survival_data}
\alias{demo_survival_data}
\title{Survival data of demo dataset.}
\format{Surv data created by Surv() function in survival package.
This data have two unnamed arguments, they will match time and event.}
\usage{
data(demo_survival_data)
}
\description{
A Surv containing survival value of 455 samples in TCGA lung Adenocarcinoma dataset.
}
\keyword{datasets}
