% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyMCB.R
\name{IdentifyMCB}
\alias{IdentifyMCB}
\title{Identification of methylation correlated blocks}
\usage{
IdentifyMCB(
  MethylationProfile,
  method = c("pearson", "spearman", "kendall")[1],
  CorrelationThreshold = 0.8,
  PositionGap = 1000,
  platform = "Illumina Methylation 450K"
)
}
\arguments{
\item{MethylationProfile}{Methylation matrix is used in the analysis.}

\item{method}{method used for calculation of correlation, should be one of "pearson","spearman","kendall". Defualt is "pearson".}

\item{CorrelationThreshold}{coef correlation threshold is used for define boundaries.}

\item{PositionGap}{CpG Gap between any two CpGs positioned CpG sites less than 1000 bp (default) will be calculated.}

\item{platform}{This parameter indicates the platform used to produce the methlyation profile.}
}
\value{
Object of class \code{list} with elements:
 \tabular{ll}{
   \code{MCBsites} \tab Character set contains all CpG sites in MCBs. \cr
   \code{MCBinformation} \tab Matrix contains the information of results. \cr
 }
}
\description{
This function is used to partition the genome into blocks of tightly co-methylated CpG sites,
Methylation correlated blocks. This function calculates Pearson correlation coefficients r^2 between
the beta values of any two CpGs r^2 < CorrelationThreshold was used to identify boundaries between any two
adjacent markers indicating uncorrelated methylation. Markers not separated by a boundary were combined into MCB. Pearson correlation coefficients between
two adjacent CpGs were calculated.
}
\details{
Currently, only illumina 450k platform is supported, the methylation profile need to convert into matrix format.
}
\examples{
demo_set<-create_demo()
#import the demo TCGA data with 10000+ CpGs site and 455 samples
res<-IdentifyMCB(demo_set)
demo_MCBinformation<-res$MCBinformation


}
\references{
Xin Yu et al. 2019 Predicting disease progression in lung adenocarcinoma patients based on methylation correlated blocks using ensemble machine learning classifiers (under review)
}
\author{
Xin Yu
}
