% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricMCB.R
\name{metricMCB}
\alias{metricMCB}
\title{Calculation of the metric matrix for Methylation Correlation Block}
\usage{
metricMCB(MCBset,training_set,Surv,testing_set,Surv.new,Method,silent)
}
\arguments{
\item{MCBset}{Methylation Correlation Block information returned by the IndentifyMCB function.}

\item{training_set}{methylation matrix used for training the model in the analysis.}

\item{Surv}{Survival function contain the survival information for training.}

\item{testing_set}{methylation matrix used in the analysis. This can be missing then training set itself will be used as testing set.}

\item{Surv.new}{Survival function contain the survival information for testing.}

\item{Method}{model used to calculate the compound values for multiple Methylation correlation blocks. Options include "svm" "cox" and "lasso". The default option is SVM method.}

\item{silent}{Ture indicates that processing information and progress bar will be shown.}
}
\value{
Object of class \code{list} with elements (XXX will be replaced with the model name you choose):
 \tabular{ll}{
   \code{MCB_XXX_matrix_training} \tab Prediction results of model for training set. \cr
   \code{MCB_XXX_matrix_test_set} \tab Prediction results of model for test set. \cr
   \code{XXX_auc_results} \tab AUC results for each model. \cr
   \code{best_XXX_model} \tab Model object for the model with best AUC. \cr
   \code{maximum_auc} \tab Maximum AUC for the whole generated models. \cr
 }
}
\description{
To enable quantitative analysis of the methylation patterns
within individual Methylation Correlation Blocks across many samples, a single metric to
define the methylated pattern of multiple CpG sites within each block.
Compound scores which calculated all CpGs within individual Methylation Correlation Blocks by SVM model
were used as the compound methylation values of Methylation Correlation Blocks.
}
\examples{
#Not run! Remove # to run the example.
 require(EnMCB)
#import exprs function
data(demo_survival_data)
datamatrix<-create_demo()
data(demo_MCBinformation)
#select MCB with at least 3 CpGs.
demo_MCBinformation<-demo_MCBinformation[demo_MCBinformation[,"CpGs_num"]>2,]

trainingset<-colnames(datamatrix) \%in\% sample(colnames(datamatrix),0.6*length(colnames(datamatrix)))
testingset<-!trainingset
#create the results using Cox regression. Remove # to run the example.
mcb_cox_res<-metricMCB(MCBset = demo_MCBinformation,
               training_set = datamatrix[,trainingset],
               Surv = demo_survival_data[trainingset],
               testing_set = datamatrix[,testingset],
               Surv.new = demo_survival_data[testingset],
               Method = "cox"
               )

}
\references{
Xin Yu et al. 2019 Predicting disease progression in lung adenocarcinoma patients based on methylation correlated blocks using ensemble machine learning classifiers (under review)
}
\author{
Xin Yu
}
\keyword{Correlation}
\keyword{Methylation}
